rename diskMaterialized
Summary:
In the fsck code, diskMaterialized does not line up with EdenFS's internal
definition of materialized. The "materialized" in diskMaterialized here is used
to mean the contents of the file are present on disk where disk means prjfs.
EdenFS may or may not think that the inode is "materialized".

Particularly, if the file is renamed EdenFS will think that the inode is
materialized, but it would not be "diskMaterialized".

This confused me and I think we should avoid the "materialized" terminology (and
probably the "disk" terminology too because materialized inodes can be "on disk"
in the overlay).

I propose the admittedly more verbose, but clear "contentsPresentInPrjfs".

Reviewed By: xavierd

Differential Revision: D41758686

fbshipit-source-id: ddd3b6eb78a9b233747c6923f01d7832cc3072ff