use Try instead of throwing
Summary:
throwing and catching exceptions is slow. This stack makes fsck 3x slower, so
in effort to reduce the cost I switched things from throwing to trys.

It didn't make much of a difference, but it is maybe a little faster. fsck use
to be ~30s and this stack makes it ~90s. Using trys saves maybe 3s as fsck
"only" took 87s after this change.

Reviewed By: xavierd

Differential Revision: D42763114

fbshipit-source-id: 2754122d340d21d613f118d124fc838e82c10bc2