add and test renamed placeholder detection
Summary:
Renamed files are treated specially in EdenFS. We materialize all renamed
files, but projfs still has these as placeholders. Generally we assume
placeholder == scm data, and full == materialized inode, but
renamed files break this rule. We need to be able to specially handle renamed
placeholders in fsck.

That's comming later in this stack. This diff adds logic to peak into the
reparse buffer and detect renames. It also comes with a full integration test
that confirms it properly detects names.

The original path is stored in the reparse buffer as well, we could use this
to detect renamed, but there seems to be a bit that indicates renames.
so it seems simpler to just use this instead.

Reviewed By: xavierd

Differential Revision: D42268211

fbshipit-source-id: 4bca09601c88297f5b33573ebe2ccad060863dc8