assert our assumptions about the reparse buffer and file attrs
Summary:
Our current FSCK implementation does not properly fsck renamed files, and to
do that we are going to need to peek inside the reparse buffer (there doesn't
seem to be anything else that indicates renamed files :( )

The format of the reparse buffer (the projFS bit) is not externally documented.
Microsoft folks have hinted at it's format on github issues:
https://github.com/microsoft/ProjFS-Managed-API/issues/55
And there are some bits with in the buffer that seem to have some clear meanings.
If we are going to rely on introspecting this buffer we should make sure we
become aware if the format changes from what we are expecting.

We also wanted to maybe limit how often we need to peek into the reparse buffer
based on file attributes. However, it does not look like that will be very
possible. To make this clear I am asserting theattributes along with the bit of
the reparse buffer we care about.

Additionally, FSCK already assumes things about what file attributes mean, and
these are kinda implementation details of ProjFS. We are exposed to those
breaking, so we should probably have tests asserting our assumptions anyways.

Reviewed By: xavierd

Differential Revision: D42766825

fbshipit-source-id: d2c022b1379e7bd57af1cd473a794607698115eb