status: fix force renamed files on EdenFS
Summary: On EdenFS when renaming a modified file into another existing file, we we showing status "?" instead of "M". In this case we get a pending change, but the renamed file has non of p1, p2 or next set in dirstate. But it does have copies, so use that to differentiate.

Reviewed By: zzl0

Differential Revision: D42998220

fbshipit-source-id: 180f7eb27c0c90b56ca1ef9af1c2f441a25980df