Remove Manifold thrift client
Summary: We used Manifold raw thrift client in the past with `walker scrub` because apparently it gave certain performance improvements over the rust client in case of small blobs (aliases and such). However, for quite some time we have been using only the rust client and the thrift version is basically dead code at this point. Manifold team recently added deprecation warnings for the thrift client which was the final push for us to get rid of it.

Reviewed By: YousefSalama

Differential Revision: D43049010

fbshipit-source-id: 0901f2a1b51799b033d4d7d3cd9e08c69a54491d