Don't use leading path sep for paths
Summary:
The paths we show in the uncommitted changes list are repo-relative files. It's kind of inaccurate to render them with a leading `/`, becuse that usually implies its at the filesystem root.

I think it makes more sense to render files without the `/` at the start.

This will also make the next stacked change to use minimal disambiguous paths work better (especially with rtl rendering used to truncate the prefix of a path instead of the suffix)

Open to feedback here. Perhaps it's useful to know when a path is root-relative?

Reviewed By: quark-zju

Differential Revision: D43058630

fbshipit-source-id: ceb69f1ad1baa96aeed591d51c1b61097892faaa