make eden starting seem less scary and hint about --wait
Summary:
1. users have mentioned this message makes it sound like eden is misbehaving,
so they try to kill it when really its just taking a while to start. We should
make the message less scary.

and

2. we have `eden status --wait` now which allows users to watch eden start
progress, so they can see it moving and feel comforted. We will probably
eventually remove this hint. Ideally we would want a hinting framework like
mercurial, so users could see it initially, and then ack it to make the
hint go away. but thats lots of work since eden doesn't have anything like that
yet. so I'm just gonna put it out there for a while.

Reviewed By: fanzeyi

Differential Revision: D43100473

fbshipit-source-id: d5518edcd2751a09e4842d619c0158562c60a20a