tunables: optionify tunables and remove get_ prefix
Summary: Currently fetching a tunable (except by repo ones) returns the default value when the tunable is not set. This can be error prune as it doesn't enforce handling the not set case and makes it awkward to set default values, it's also not consistent with the by repo tunables which return an option instead. This diff changes tunable fetching to make it return an option, and adds a .unwrap_or_default() at every callsite to maintain the current behaviour.

Reviewed By: yancouto

Differential Revision: D43123305

fbshipit-source-id: 32cb69ba78d52bc22979250c6e16edd1800cd1de