Add tracking for client connection
Summary:
Add tracking for client connection to mark when a new ISL window is opened. This applies for all platforms since `onClientConnection` is the server-side entry point.

I include the cwd here since it's not part of the scuba columns. This lets us have a way of tracking the cwd per session.

Reviewed By: quark-zju

Differential Revision: D43164387

fbshipit-source-id: 1b162abefcffe52b314ec6f7047a0b11192bb21d