pathutil: follow symlinks outside repo to inside repo
Summary:
In general we don't follow symlinks since symlinks can be under version control, and it is a vector to reveal file contents outside the repo accidentally.

However, it's relatively common for users to have symlinks in their home directory that points somewhere inside the repo, for example ~/myproj points to ~/somerepo/some/long/path/myproj. Previously running commands like "hg revert ~/myproj/foo" wouldn't work because ~/myprof/foo appears to not be under ~/somerepo (and we don't follow symlinks).

With this change, we now follow symlinks one-at-a-time. For example, in the above case we would resolve "~/myproj" to "~/somerepo/some/long/path/myproj" and then the "revert" allowing the revert to work. Note that we stop following symlinks as soon as we are under "root". This is to avoid following symlinks inside the repo.

There was an existing test explicitly showing this new behavior doesn't work. It wasn't clear why that test was added, but I can't think of a reason why this behavior would be unsafe.

Reviewed By: zzl0

Differential Revision: D42929869

fbshipit-source-id: 0a8293d82ccfbf7c6903aba9e3a7052f59b57b74