Add Collapsable component
Summary:
Add a simple component that lets you click to collapse/expand it.

Note: we probably need to be a bit more careful about accessibility for a component like this. Currently, this is only used in one minor place but it's worth expanding this to be better about keyboard navigability and aria roles.

Reviewed By: quark-zju

Differential Revision: D43133033

fbshipit-source-id: adffd7ba1730f8f011b4486e904dfc0e8e71833c