Support explicitly dispatching commands
Summary: Our keyboard shortcut system lets us represent shortcuts as "Commands", which can be dispatched with shortcuts. It's useful to also be able to dispatch commands explicitly as a callback in code. For example, you can trigger `Escape` key presses to dismiss things with long-range impact, and the code is very readable.

Reviewed By: quark-zju

Differential Revision: D43104460

fbshipit-source-id: e2606270983f9365e2fa106250daa949f14f2229