util: drop ui.threaded support
Summary:
`ui.threaded` was added to run main Python logic in another thread so Ctrl+C
can still be handled in the Python main thread. It is only used in tests and
does not play well with native code in general (ex. changes how
mysql-connector behaves). Now we want to use Rust to handle Ctrl+C. So let's
drop the incompatible Python logic.

Reviewed By: muirdm

Differential Revision: D42909480

fbshipit-source-id: 082ba45680209dea6dee1b8f764bcb0dd364246b