sqlitecatalog: remove O(N^2) in WindowsFSCK
Summary:
In FSCK, the repository is crawled recursively from the root, and for each
directory, the ObjectStore is queried to obtain a Tree for that directory. The
way the ObjectStore is queried is by starting from the root and iterating over
all the path component, unfortunately, this means that the ObjectStore is
queried in a quadratic fashion. This can be avoided by recognizing that the
FSCK code is already aware of the parent Tree, and thus we can simply query the
ObjectStore for the last path component.

Reviewed By: kmancini

Differential Revision: D43008298

fbshipit-source-id: 143e0f93e566d17d19304a8200c3ed15d7a3b0c7