Expose repo-name encoding scheme for sharded clients
Summary: When talking to sharded services through SR, the `domain` name needs to be provided which roughly maps to a repo in Mononoke services. However, SM is not as permissive as Mononoke when it comes to naming schemes (i.e. `/` and `+` are not allowed) which is why we use repo-name encoding-decoding. This diff exposes that functionality so it can be used outside of this library.

Differential Revision: D43185709

fbshipit-source-id: 2770586b013beee07d9ac65242516c3e1c4ea0d5