status: use matcher when finding NEED_CHECK files
Summary:
This way we skip ignored files that are being tracked in the dirstate. Previously we were comparing the contents of the ignored files only for them to be excluded later. It also may help in a similar fashion for sparse repos w/ a lot of files outside sparse profile present in the working copy.

This only applies for watchman repos, not EdenFS. EdenFS only tracks newly added or removed files in the dirstate.

Reviewed By: quark-zju

Differential Revision: D43141286

fbshipit-source-id: 82617f9b37a58b1b9181c1dcfb12ade06e6cd603