status: handle non-utf8 paths from watchman
Summary:
We don't intend to fully support non-utf8 file names, but it is important we don't error out.

We now use the new BytesNameField watchman query type so we can decode non-utf8 file paths in watchman responses. Then, we pipe the invalid paths along up to the status command where they are printed out as warnings to the user.

In the future we probably want to track invalid paths in the treestate and make "hg clean" support removing them as a special case (as suggested by quark-zju).

Reviewed By: quark-zju

Differential Revision: D43141278

fbshipit-source-id: ad04ebdf75594e33ebaded137e9c65c487f0b30b