Update useModal to allow a primary button
Summary:
Update `useModal` with a `type: 'confirm'` to allow taking object configs for buttons, instead of just strings.

So you could use it like this:
```
function MyComponent() {
  const showModal = useModal();
  return <button onClick={() =>{
    const response = await showModal({
      type: 'confirm',
      title: 'Do the thing?',
      buttons: [
        {label: 'Confirm!', primary: true},
        {label: 'Cancel'}
      ]
    });
  }}>click me</button>
}

```
And the "Confirm!" button would be blue and focused by default.

In the future we could add additional features like tooltips to this button config. It's an API I've seen in other places. But it's also nice to support just plain strings too.

Reviewed By: quark-zju

Differential Revision: D43200834

fbshipit-source-id: 9fdb7429e5012f5c98728c0edee1c5f59ddd9b22