commit_graph_compat: start bubbling up errors when adding to the new commit graph
Summary: The only type of errors found in mononoke_commit_graph scuba table were SQL errors, and not bubbling up such errors causes the commit graph db tables to drift out of sync with the changesets table over time. This diff also removes the timeout as the only times we hit it are either inserting a large number of changesets or sql connection timeouts.

Reviewed By: RajivTS

Differential Revision: D43124401

fbshipit-source-id: 35c9995d73fb18b52475b50c741fcf726b3e9ae6