Tweak styling of Load More, render as ellipsis on graph
Summary:
After dogfooding the "load more commits" button for a few days, I found it was sticking out a bit more than I like. I tried making it work as a vertical ellipsis on the commit graph and I liked how that looked. This makes the top of the commit graph render the same as the bottom of the commit graph, so they both have ellipses. The fold icon was pretty accurate, but I think vertical ... is also pretty sensibe for loading more, especially when it renders as part of the branch.

To do this, we need to be a little sneaky in our rendering. I put the ellipsis icon as a child of the commit tree, then shift it with negative margin, and put a background behind it so it covers the branch line. This fixes an issue that made `<MainLineEllipsis />` on the bottom of the tree not render at all.

curious if you have thoughts sggutier, since you added this originally!

Reviewed By: sggutier

Differential Revision: D43249895

fbshipit-source-id: 481e9762386aadde3ce865ffc3fbd1cd4f315e46