Rename prepushrebase to in_critical_section and do it in the critical section
Summary:
This fixes a current bug in our landing logic: The `prepushrebase` call happens concurrently with fetching the bookmark value, which means we can't guarantee no more commits land before actually starting this pushrebase.

But some pushrebase hooks, like globalrev, depend on that. Because they calculate the `max_log_id`, which will change if more commits land.

This can lead to a race condition where we try to set the same globalrev for a commit. In our code, we fail when that happens, but if we're a little less careful it could cause inconsistency issues.

So let's make things clear by making it always happen after the critical section starts, and making that clear in the API.

Reviewed By: markbt

Differential Revision: D42805082

fbshipit-source-id: 78043761cb33f8ead30b2685d5e9983fcdfbedc4