Require RepoConfig on cross_repo_sync::Repo and do some deblobrepoification
Summary:
This adds the requirement that `cross_repo_sync::Repo` trait implements `RepoConfig`. This is being used on `repo_import`, and will be used on the backsyncer in D43004874.

Since `BlobRepo` doesn't have `RepoConfig`, this causes all sorts of problems, and I had to do some deblorepoification to be able to pass this argument along, which involved refactoring a bunch of tests and classes.

Reviewed By: mitrandir77

Differential Revision: D43085593

fbshipit-source-id: 642286731b5bbb7017fcf315c98cd92f6e3048c0