use Rust implementations in integration tests
Summary:
In production, we read /etc/edenfsctl_rollout.json (or an equiv on Windows) to determine whether to use the Rust or Python version of subcommands. Chef writes this file and edenfsctl wrapper reads it if it exists.

If edenfsctl_rollout.json does not exist, then we use a hardcoded list of "experimental" commands to determine whether we should run Python or Rust. This hardcoded list is defined in Rust.

When running integration tests, the file that Chef creates does not exist at the custom --etc-eden-dir location. For integration tests, we should create this file at our custom --etc-eden-dir location and leave it empty so that we will default to Rust for testing all subcommands.

This can be extended in the future so that we can pass a custom list of subcommands to test w/ Python.

Reviewed By: fanzeyi

Differential Revision: D42684141

fbshipit-source-id: d13b414239cc554e41680f33ea4a6a58b3f0b78f