read config to determine whether to use dmg or apfs bind redirections
Summary:
On macOS, we can either use disk images or APFS subvolues for bind-mount redirections. We currently default to APFS subvolumes.

We'd like to try moving back to disk images because APFS subvolumes have many limitations that we didn't know about prior to using them. In particular, macOS Sandcastle hosts are limited to a maximum of 20 APFS subvolumes. This causes problems when we try to run redirect integration tests on macOS.

*This diff updates the Rust redirect logic so that we can utilize dmg images with the flip of a config option. The config option currently defaults to "apfs".*

Reviewed By: fanzeyi

Differential Revision: D42371702

fbshipit-source-id: 7afafffbf2be84bee17287c12656f67b58678bb0