add etc eden dir flag to daemon invocation of `redirect fixup`
Summary:
`eden redirect` reads config options to determine which redirection type to use on macOS. Therefore, we need to pass proper information about where to find the configs (in this case, etc_eden_dir) to edenfsctl when the daemon invokes it.

If we don't do this, our integration tests read production configs because edenfsctl defaults to using the default etc_eden_dir location instead of whichever etc_eden_dir location was passed to the test daemon when it was started.

Reviewed By: mitrandir77

Differential Revision: D42560215

fbshipit-source-id: 086b6a709c9c9563b67c06c095e97c205e0c998e