listen to PRE_CONVERT_TO_FULL
Summary:
Andrew Krieger reported that eden doesn't seem to notice when files are
truncated on windows. And it repros pretty easily:

```
fbclone fbsource tmpfbsource
cd tmpfbsource/arvr/libraries/avatar/Libraries/asset/items/gltf
Clear-Content src/GltfCompactSkinningData.cpp
hg st
# empty
```

Eden is getting a file closed and NOT modified notification. Microsoft has
confirmed that this is unexpected and would fix it.

But in the meantime there is something we can do to work around it. We do get
the notification "pre convert to full". Full files should be materialized in
EdenFS, so we can use this notification to mark the inode as materialized early.

Handing notifications is async, so this should not effect performance.
Benchmarking creating a lot of files, shows that this doesn't seem to make much
of a difference.

We will roll this out with a config, and then roll it back once microsoft's fix
rolls out.

Reviewed By: xavierd

Differential Revision: D42465547

fbshipit-source-id: cc41fa8c7ad7fbf96ae4a9af5b3a7e6f477fc449