Fix Mononoke crashes due to Manifold UAF issue
Summary:
Initially reported in Mononoke server, this issue turned out to be quite more wide-spread, impacting many more of our tools and services. On a high level, the problem arises when we have multiple inflight get/put/is_present request and wait for only a quorum of them. The exact number of operations we wait for depends on the blobstore config and the structure of the multiplex. The remaining blobstore ops don't get polled once they are fired off.

I tried to counter this through cancellation token where blobstore ops would be `selectively` waited for along with a cancellation future. If the cancellation is signalled, the other future (i.e. blobstore ops) would be aborted and a no-op result would be returned. The signalling of the cancellation was done in `impl Drop` before we exit. However, this worked only for the components executing under the tokio runtime. Manifold's rust-cxx library also involves folly which doesn't interpret this cancellation as expected.

There are three options available at this point.
- Communicate the cancellation to the C++ async part and make the relevant code honor this cancellation.
- Conditionally wait in `impl Drop` of the multiplex
- Always wait in places where this blows up

I do not know if the first option can even be done reliably without introducing new bugs. The second option is local to our codebase and its a one-place change that would fix this problem pretty much everywhere in Mononoke.

Reviewed By: YousefSalama

Differential Revision: D43193118

fbshipit-source-id: 2da3f3228ea948eaaa91aebdb3174ce5a5a1d255