pr submit: don't update closed PRs
Summary:
If a PR is not open (i.e. merged or closed), "sl pr submit" will now avoid updating it. Previously it didn't consider the PR's state.

We emit a warning and a hint pointing the user to "sl pr unlink" if they want to create a new PR. (After disassociating the commit from the existing PR, "sl pr submit" will create a new PR).

Fixes #474

Reviewed By: quark-zju

Differential Revision: D43208767

fbshipit-source-id: 964381a6fcb9de30b2751bb825509cca2fd68f69