fix `redirect fixup` UNC path errors on Windows
Summary: std::canonicalize returns a UNC path on Windows. This causes all sorts of issues when we try to use this UNC path elsewhere. We should strip the UNC prefix before trying to pass the path to buck commands. This will prevent the "illegal characters in path" error that users were seeing (see this paste: P623150778).

Reviewed By: kmancini

Differential Revision: D43293847

fbshipit-source-id: 0289c288c180ca1fe34b59babb3f13740c919600