add counter for filesystem sync duration
Summary:
Currently EdenFs does not realize that truncated files have changed because
PrjFS sends a "file closed no modification" notification instead of a
"file closed and modified" notification.

Microsoft is fixing this, but in the meantime to work around this we are
listening to a new notification "file about to be converted to a full file".

This notification is sent anytime a placeholder is converted to a full file.
So we are going to be generally listening to more notifications. Notifications
are handled on a background thread, so there should not be a big filesystem io
perfomance hit from this.

However, we do wait to process all notifications before we answer thrift
requests. more notifications could mena this slows down. so I am adding a counter
here to be able to directly monitor this.

Mike's rolling out ODS counter collection on windows currently (D43198806), so
we should be able to monitor this to make sure we are not taking to severe a
performance hit.

Reviewed By: MichaelCuevas

Differential Revision: D43286103

fbshipit-source-id: 46a10c0ed1f1447bff48a9f82fc092939ebdaab0