Add tracking when rendering an empty state
Summary:
Add some track calls for empty state rendering. These represent errors / invalid states where the repository could not be successfully created. We use different track event names for each empty state cause, so we can measure which cases are happening and when.

Also fixing a bug here where we weren't calling track when the repository is not hooked up.

Reviewed By: quark-zju

Differential Revision: D43292235

fbshipit-source-id: d5b3e0b3a299826ba3b9766c5ade3a545c86bd22