admin: Verify bookmarks should use commit sync outcome
Summary:
This command was not working because it always needs a remapping of the commit, but on this case it was a `EquivalentWorkingCopyAncestor`.

This is probably out of data, as it uses a bunch of manual "CommitSyncer"s instead of using `Syncers` which have both directions and are easier to use.

Reviewed By: mitrandir77

Differential Revision: D43047554

fbshipit-source-id: 5f1050c0009355c735ceba93b3da9697661230b1