github: run "mark pr landed" inline instead of as hook
Summary:
Hooks are annoying since it is hard to control which "sl" binary is used to invoke the hook. Instead, let's just call the landed pr hook inline using the classic extension wrapping technique.

I made a lazy decision to disable the hide-landed-commits logic by default for tests since the additional GH activity requires tests to mock more stuff out that they aren't necessarily testing.

I also moved the enablement of the "github" extension from the sapling config (which is activated when $0 is "sl") to the git config (which is activated when in a git repo). It doesn't seem like GH integration needs to be coupled with the Sapling rebrand.

Fixes #457

Reviewed By: quark-zju

Differential Revision: D43276199

fbshipit-source-id: 3feb7621a037190044c41020e6a965f8b62bc432