sparse: filter excluded files from "status" output
Summary:
Previously if the dirstate and working copy were out of sync with regard to the sparse profile, the Python "sl status" would report modified files even if they were excluded in the sparse profile. Fix by applying the sparse matcher in dirstate.status().

This fixes "sl status" to be consistent with "sl commit", "sl diff", etc. in that it filters out files not in the sparse profile.

Note that this does not fix the reverse case where a file was added to .hg/sparse without going through "hg". The file won't be tracked until "hg sparse include" or "hg sparse refresh" is called. Fixing this is more involved, but Jun suggests storing the sparse profile in the dirstate metadata so status can detect out-of-band sparse changes and automatically refresh the (minimal) differences.

Reviewed By: zzl0

Differential Revision: D43256615

fbshipit-source-id: fda28d4fb833ec36c2d57f3be626e325da3be654