gnerate bundles concurrently to sending them
Summary:
Currently the mononoke_hg_sync_job either prepares a bundle or sends it to hg
servers. Those steps could be done concurrently. This dfff creates a stream of
bundles that is buffered so the syncing to hg servers doesn't have to wait for
new bundle generation.

The implementation is a bit gross, unfortunately I couldn't get the lifetimes
arround `scan()` transofrmer to work without mutex and I hit higher-order
lifetime exception few times which caused me to wrap some things in Arc.

Reviewed By: yancouto

Differential Revision: D43306918

fbshipit-source-id: b2ec3da21f4791c57a02fcfe76c735f4fdb16efd