[sapling] Remove Cargo.lock on `make clean`
Summary:
The help text for `make clean` says:

> remove files created by other targets
> (except installed files or dist source tarball)

So I'd expect it to remove any non-installed built state like Cargo.toml.

Concretely, I was surprised by this behavior when the build failed after
pulling in updates, even after running `gmake clean`.  My Cargo.lock from a
previous build had pinned the watchman_client crate to an old version, but the
updated workingcopy crate relied on the addition of `BytesNameField` to
watchman_client@main.

I think it would be better if autocargo generated a Cargo.lock for us from the
Buck versions of dependencies, but as long as it's being generated as a build
artifact it should be removed by `make clean`.

Test Plan:
```
% cd eden/scm
% gmake oss
% ls Cargo.lock
Cargo.lock
% gmake clean
% ls Cargo.lock
% ls: Cargo.lock: No such file or directory
```