Add object exists check in create_git_tree method
Summary: The `create_git_tree` method currently converts the given git tree hash into a Mononoke `BonsaiChangeset` and stores it in the data store. An implied pre-requisite in this case is the corresponding raw git tree object should already be present in the blobstore. However, this invariant might be violated. This diff modifies the `create_git_tree` method to check for the presence of the git object before converting it into BonsaiChangeset

Reviewed By: mitrandir77

Differential Revision: D43363613

fbshipit-source-id: 8a390bdc86e7a5e857d8a0d5e5cf4dcb34e50402