Implement create_git_tree method in Mononoke API
Summary: This diff implements the `create_git_tree` method which is responsible for creating Mononoke counterpart of the git tree object. Note that we do not create a new `bonsai_git` mapping since that is just for actual commits. Additionally, we do not log this new changeset to the scribe category since we don't want the commit tailers to be notified of this change. Except the `git_tree_hash` property, rest of the changeset properties get default values.

Differential Revision: D43306786

fbshipit-source-id: 850bf2dae8aaf525348d1a27a5ca73c51edb7af9