repo: do not construct EdenApi if remote is not configured
Summary:
While `EdenApi` is used in production use-cases. There are a lot of cases in
tests (ex. `newclientrepo`) that the EdenApi should not be constructed. For
example, if the repo does not have a remote repo configured (path.default is
empty) then EdenApi should not be used.

In Python `repo.nullableedenapi` is used to avoid creating EdenApi in such
cases. However, the Rust code does not have such logic. This diff adds such
logic in Rust so it understands when not to create EdenApi.

Reviewed By: muirdm

Differential Revision: D43263816

fbshipit-source-id: acf79c1d9988f8760e34700874c92c6ae8d3f057