revisionstore: add an API to get aux data from StoreTree
Summary:
On EdenFS, it is very common to get a Tree and then immediately after read its
aux data. This is so common that the lock on the aux data IndexedLog is
extremely contended, leading to most of the time being spent waiting on taking
it. To avoid this, we should expose the aux data already present in the EdenAPI
result and return it to EdenFS immediately.

For now, this merely introduce the API, future diffs will plug it deeper into
EdenFS.

Reviewed By: muirdm

Differential Revision: D43132763

fbshipit-source-id: 13d2158e6f2b6d52905ce684bc0f12096a95920c