pass -A flag to commit/amend to addremove
Summary:
Commit/amend fail if you pass in untracked files. If we pass `--addremove` these commands will succeed.

 The CLI works by using tracked files as a way of specifying which files you want to include. But the UI is not really limited to such a system. We already have a checkmark system next to files you're going to commit/amend. That already represents the idea of "tracking" this file or not. It's purely extra overhead to think about if the files are actually tracked by `sl` or not.

In a subsequent diff we'll need to change how untracked files are rendered. No reason to show them as a question mark if it doesn't help you to make any decisions on your actions in the UI.

Reviewed By: quark-zju

Differential Revision: D43205299

fbshipit-source-id: 22cf28ec2ae55ec9ec599763597ddde818e5200a