Add InlineErrorBadge and improve ErrorNotice
Summary:
Add a new component to render short inline error messages. In general, I think it's nice to show contextual, inline errors when they occur (as opposed to a global error state). So when uploading an image fails, we should show the error directly in the uploading area. This inline element makes that easy to express.

The inline error badge can also show the error in question when hovering, so it's easy to dig into details when you hit an issue.

While styling this, I found I liked this as a lighter red background with darker text rather than a red border. I figured for consistency it would make sense to update `<ErrorNotice>` at the same time with a similar look. I like how the new error notice looks a lot more, it's less harsh.

Reviewed By: quark-zju

Differential Revision: D43411125

fbshipit-source-id: ae2c503f3e8c2f8e075cdf2fdd99d9a5bc6b19d2