Add nextMessageMatching util
Summary:
Add a util that lets you extract one message from the stream of messages. You provide a message type and a predicate to check for. It returns a Promise that resolves when it sees a message that matches, then disposes its listener.

This is useful for "subscribing" to look for a message with a matching ID.

Reviewed By: quark-zju

Differential Revision: D43378566

fbshipit-source-id: 039641ad530fa6dc57b25f042a475fe962a91ea5