Add file picker to text field
Summary:
See Previous diff for context about image uploading.

If a user doesn't know about drag&drop and pasting images, it's hard to learn about. By having an explicit *click to upload file* UX, it becomes discoverable. Also some users may struggle with drag&drop but can handle a click-to-upload.

(TODO: we should handle keyboard-only opening of the file picker for accessibility)

In this diff I add the notion of the EditorToolbar for a CommitInfoField. This is a button bar floating at the bottom of the text area that can include various buttons/editor status/messages. This is a common UI pattern for adding features to a text area.

Here we add the file upload button and later add the status of ongoing uploads and let you cancel ongoing uploads.

In the future we may add additional buttons to this toolbar, like filling from a template or whatever

The file picker itself is an icon button using a paperclip icon. That's the typical icon used for attaching files (email, facebook composer, etc). Surprisingly vscode codicons doesn't have a paperclip icon, so I had to make one myself. I styled it with width that roughly matches codicons. I might make a PR to codicons to add this icon.

The file picker is handled as an `<input type="file" />`. To customize rendering, we need to use a `<label>` & hide the input, so clicking on the label opens the input even though it's hidden.

Reviewed By: quark-zju

Differential Revision: D43332597

fbshipit-source-id: bfce7e9a866c2993e6280d0054c7751c8f366179