Add tests for image uploading
Summary:
Actually add tests for all the image uploading features added in this stack.

We need to hardcode image uploading to be enabled for our tests, since it may be disabled in production / OSS builds.

These tests go over pasting/dragging/choosing files, making sure it sends the data to the server, and the placeholder&replacement system for the final uploaded links

Reviewed By: quark-zju

Differential Revision: D43408510

fbshipit-source-id: d581bb1df794e09bef15d944b52a01659f74bfb2