Refactor utils for testing commit info view
Summary:
We had a bunch of small utils nested inside CommitInfoView.test.tsx. This test file is getting quite large, so I want to put new tests that relate to things inside the commit info view into a separate file. A nice first step is to extract these utils, so we can get the same ease of testing in other files as well.

inside this test file, it's implicit that we are always acting on the CommitInfoView, so we can destructure these utils at the top level so you don't need to qualify your uses with `CommitInfoTestUtils`. Other files may want to use the full name for clarity

Reviewed By: quark-zju

Differential Revision: D43408512

fbshipit-source-id: fbf2bb6348c809d001236aa4318e1f093acbf939