Send messages with binary payloads
Summary:
To support uploading images to test plans, we need to send the file data over to the server.

We do this by sending ArrayBuffers to the server along the websocket.
Other ideas don't work as well:
- sending as text via base64 encoding is less efficient
- sending filenames so the server can read the file itself doesn't work because the server may be running on a different host than the client

ArrayBuffers are supported in websocket messages. However, they cannot be embedded into our JSON messages. Doing so would make the entire message binary and so JSON parsing would fail.

We also need to send context along with the binary data (what its for, metadata about the file, etc), so it doesn't fully suffice to just send the ArrayBuffer directly.

Instead, the idea is that we send two messages back-to-back.  The client's API is to send a JSON message with a binary payload as a second argument. First, we send the JSON message that declares that it expects a binary message to follow it (plus any metadata, filename, how we expect to use it, etc). Then we send the binary ArrayBuffer.

We then reconstruct a message+payload on the server, and can put it into another `switch(message.type)` to handle it. This lets us support multiple kinds of binary messages in the future.

Note that websockets are a reliable transport, so we can guarantee messages will come in-order, so we never expect the message&payload to not come sequentially. If we ever saw that, it would be a programming error on our part and not a networking bug.

The use case at hand is uploading images to embed links. This diff includes the `uploadFile` type to kickstart the use case of sending binary messages. We'll need to include more info this message type in the near future: file name, file type, unique ID, etc.

Reviewed By: quark-zju

Differential Revision: D43369819

fbshipit-source-id: 59dcd03b2b891f54aa12f91b25fadfe292ce67d4