Support drag&drop image files onto text area
Summary:
Allow dragging an image file onto the text area in the commit info view. This will upload the image to the code-review-provider-specific image uploading server, and then insert the link to the file where your cursor was.

For Github, this will upload to github like if you dragged an image into a Pull Request editing area. And we have a separate image upload service internally.

This diff does not handle the file uploading itself. Instead, it just waits for 30s before returning the file name.

later in the stack, we'll add the rest of the infra to actually do the uploads of the files. The first few diffs in this stack will be setting up the mechanisms to get the files we want to upload and making the UX for it work.

This diff only adds support for drag&drop, but we'll also have a file picker and copy/paste support in the next diffs. Plus cancelling, progress, etc.

Drag & drop support is not too difficult, we mostly just need an `onDrop` callback on a wrapper div. We use onDragEnter/Leave to set a highlight class which lets us render a highlight over the entire droppable area, plus text prompting the drop.

More of the complexity in this diff is supporting the file upload state. Namely, how we show ongoing uploads and insert the resulting short link.

We want to insert the file where your cursor is. This comes from experience in the previous ISL implementation. It's not very nice to drop the link at the end of the text area.

If inserting at the cursor, since the upload takes time, we need to track that cursor location. Also from experience in the previous ISL, it's not as nice to drop the link whereever your cursor ends up once the file is ready.

A solution that gives a nice-feeling UX is to drop placeholder text where the cursor is, and replace the placeholder with the link once the upload succeeds. This keeps the link in the place you initially wanted.

The downside of this approach is that the user could possibly delete some of the text of the placeholder, preventing future replaces. This isn't a huge deal, but a better approach might be to use a contenteditable div so we can insert an HTML spinner that we can find&replace rather than literal text. Unfortunately contenteditable has its own quirks so I'd rather test out such a change later on.

We also need to be slightly careful about replacing the placeholder, so we preserve your selection. And lastly, we want to selectively insert spaces around the inserted text so that the link doesn't collide with other characters that make it an invalid link.

Reviewed By: quark-zju

Differential Revision: D43332600

fbshipit-source-id: e7453084f15e46caefb7a6a4c8d62e5f0419b85c